+++ fkie_mas_daemon/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package fkie_mas_daemon
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.5 (2024-07-24)
+------------------
+* fkie_mas_daemon: added dynamic-reconfigure.py script
+* Contributors: Alexander Tiderko
+
 3.0.4 (2024-07-19)
 ------------------
 * fkie_mas_daemon: changed return type of ros.launch.changed uri